app/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0
python-dotenv>=1.0.0
httpx>=0.27.0
pyserial>=3.5
itsdangerous>=2.1.0
python-multipart>=0.0.9
jinja2>=3.1.0
